+++ Kernel/System/TemplateGenerator.pm
@@ -1118,6 +1118,12 @@ sub _Replace {
     # cleanup
     $Param{Text} =~ s/$Tag.+?$End/-/gi;
 
+    my $HashGlobalReplace = sub {
+        my ($Tag, %H) = @_;
+        my $Keys = join '|', map { quotemeta } grep { defined $H{$_} } keys %H;
+        $Param{Text} =~ s/(?:$Tag)($Keys)$End/$H{$1}/ig;
+    };
+
     # get owner data and replace it with <OTRS_RESPONSIBLE_...
     $Tag = $Start . 'OTRS_RESPONSIBLE_';
     if ( $Ticket{ResponsibleID} ) {
@@ -1137,10 +1143,7 @@ sub _Replace {
         }
 
         # replace it
-        for ( sort keys %Responsible ) {
-            next if !defined $Responsible{$_};
-            $Param{Text} =~ s/$Tag$_$End/$Responsible{$_}/gi;
-        }
+        $HashGlobalReplace->($Tag, %Responsible);
     }
 
     # cleanup
@@ -1163,12 +1166,7 @@ sub _Replace {
         }
     }
 
-    # replace it
-    for ( sort keys %CurrentUser ) {
-        next if !defined $CurrentUser{$_};
-        $Param{Text} =~ s/$Tag$_$End/$CurrentUser{$_}/gi;
-        $Param{Text} =~ s/$Tag2$_$End/$CurrentUser{$_}/gi;
-    }
+    $HashGlobalReplace->("$Tag|$Tag2", %CurrentUser);
 
     # replace other needed stuff
     $Param{Text} =~ s/$Start OTRS_FIRST_NAME $End/$CurrentUser{UserFirstname}/gxms;
@@ -1260,14 +1258,7 @@ sub _Replace {
     }
 
     # replace it
-    for ( sort keys %Ticket ) {
-        next if !defined $Ticket{$_};
-        $Param{Text} =~ s/$Tag$_$End/$Ticket{$_}/gi;
-    }
-    for ( sort keys %DynamicFieldDisplayValues ) {
-        next if !defined $DynamicFieldDisplayValues{$_};
-        $Param{Text} =~ s/$Tag$_$End/$DynamicFieldDisplayValues{$_}/gi;
-    }
+    $HashGlobalReplace->($Tag, %Ticket, %DynamicFieldDisplayValues);
 
     # COMPAT
     $Param{Text} =~ s/$Start OTRS_TICKET_ID $End/$Ticket{TicketID}/gixms;
@@ -1303,10 +1294,7 @@ sub _Replace {
 
         # replace <OTRS_CUSTOMER_*> tags
         $Tag = $Start . 'OTRS_CUSTOMER_';
-        for ( sort keys %Data ) {
-            next if !defined $Data{$_};
-            $Param{Text} =~ s/$Tag$_$End/$Data{$_}/gi;
-        }
+        $HashGlobalReplace->($Tag, %Data);
 
         # replace <OTRS_CUSTOMER_BODY> and <OTRS_COMMENT> tags
         for my $Key (qw(OTRS_CUSTOMER_BODY OTRS_COMMENT)) {
@@ -1477,16 +1465,11 @@ sub _Replace {
         }
 
         # replace it
-        for my $Key ( sort keys %CustomerUser ) {
-            next if !defined $CustomerUser{$Key};
-            $Param{Text} =~ s/$Tag$Key$End/$CustomerUser{$Key}/gi;
-            $Param{Text} =~ s/$Tag2$Key$End/$CustomerUser{$Key}/gi;
-        }
+        $HashGlobalReplace->("$Tag|$Tag2", %CustomerUser);
     }
 
     # cleanup all not needed <OTRS_CUSTOMER_DATA_ tags
-    $Param{Text} =~ s/$Tag.+?$End/-/gi;
-    $Param{Text} =~ s/$Tag2.+?$End/-/gi;
+    $Param{Text} =~ s/(?:$Tag|$Tag2).+?$End/-/gi;
 
     return $Param{Text};
 }